Add jwt to application services

+++ Services/PostService.cs
@@ -1,8 +1,10 @@
 using Api.Contracts;
 using Api.Contracts.ServiceContracts;
 using Api.Entities.Dtos;
+using Api.Entities.Exceptions;
 using Api.Entities.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Identity;
 
 namespace Api.Services;
 
@@ -10,11 +12,13 @@ public class PostService : IPostService
 {
     private readonly IRepositoryManager _repositoryManager;
     private readonly IMapper _mapper;
+    private readonly UserManager<ApplicationUser> _userManager;
 
-    public PostService(IRepositoryManager repositoryManager, IMapper mapper)
+    public PostService(IRepositoryManager repositoryManager, IMapper mapper, UserManager<ApplicationUser> userManager)
     {
         _repositoryManager = repositoryManager;
         _mapper = mapper;
+        _userManager = userManager;
     }
 
     public IEnumerable<Post> GetAllPost(bool trackChanges)
@@ -23,28 +27,41 @@ public class PostService : IPostService
         return posts;
     }
 
-    public IEnumerable<PostDto> GetUserPosts(string userId, bool trackChanges)
+    public async Task<IEnumerable<PostDto>> GetUserPosts(string userId, bool trackChanges)
     {
+        await FindUser(userId);
         var posts = _repositoryManager.PostRepository.GetUserPosts(userId, trackChanges);
         return _mapper.Map<IEnumerable<PostDto>>(posts);
     }
 
-    public Post? GetPost(string userId, int postId, bool trackChanges)
+    private async Task FindUser(string userId)
     {
-        var post = _repositoryManager.PostRepository.GetPost(userId, postId, trackChanges);
-        return post;
+        var user = await _userManager.FindByIdAsync(userId);
+        if (user is null) throw new UserNotFoundException($"User with id {userId} not found");
     }
 
-    public Post CreatePost(string userId, CreatePostDto createPostDto)
+    public async Task<PostDto?> GetPost(string userId, int postId, bool trackChanges)
     {
+        await FindUser(userId);
+        var post = _repositoryManager.PostRepository.GetPost(userId, postId, trackChanges);
+        if (post is null) throw new PostNotFound($"Post with id {postId} not found in database");
+        return _mapper.Map<PostDto>(post);
+    }
+
+    public async Task<PostDto> CreatePost(string userId, CreatePostDto createPostDto)
+    {
+        await FindUser(userId);
         var post = _mapper.Map<Post>(createPostDto);
         _repositoryManager.PostRepository.CreatePost(userId, post);
         _repositoryManager.Save();
-        return post;
+        return _mapper.Map<PostDto>(post);
     }
 
-    public void DeletePost(Post post)
+    public async Task DeletePost(string userId, int postId)
     {
+        await FindUser(userId);
+        var post = _repositoryManager.PostRepository.GetPost(userId, postId, true);
+        if (post is null) throw new PostNotFound($"Post with id {postId} not found in database");
         _repositoryManager.PostRepository.DeletePost(post);
         _repositoryManager.Save();
     }
